# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "AI safety proxy that screens every request and response between client applications and Google Vertex AI.\nSupports Anthropic Messages, OpenAI Chat Completions, and Vertex AI Gemini native formats.",
        "title": "Bulwarkai Proxy Service",
        "contact": {
            "name": "Dave Walker",
            "url": "https://github.com/Daviey/bulwarkai"
        },
        "license": {
            "name": "Apache-2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/v1/chat/completions": {
            "post": {
                "description": "Proxies OpenAI-format requests to Vertex AI Gemini, screening prompt and response",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OpenAI"
                ],
                "summary": "OpenAI Chat Completions API",
                "parameters": [
                    {
                        "description": "OpenAI Chat Completions request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "invalid json",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v1/messages": {
            "post": {
                "description": "Proxies Anthropic-format requests to Vertex AI Gemini, screening prompt and response",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Anthropic"
                ],
                "summary": "Anthropic Messages API",
                "parameters": [
                    {
                        "description": "Anthropic Messages request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "invalid json",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKey": {
            "type": "apiKey",
            "name": "X-Api-Key",
            "in": "header"
        },
        "BearerToken": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "ForwardedAccessToken": {
            "type": "apiKey",
            "name": "X-Forwarded-Access-Token",
            "in": "header"
        }
    }
}